+++ src/actions/meetingsFetch.test.js
@@ -1,22 +1,4 @@
-import { loginFailure, meetingsFetchFailed } from './index';
-
-describe('Login failure action creator', () => {
-  const error = new Error('Oh no!');
-
-  it('to create a properly formed action.', () => {
-    expect(loginFailure(error)).toEqual({
-      type: 'LOGIN_FAILURE',
-      payload: error,
-      error: true,
-    });
-  });
-
-  it('to return an error message.', () => {
-    const action = loginFailure(error);
-    const errorMessage = action.payload.message;
-    expect(errorMessage).toEqual('Oh no!');
-  });
-});
+import { meetingsFetchFailed } from './index';
 
 describe('Meeting fetch action creator', () => {
   const error = new Error('Oh no! Failed to fetch meetings.');
